Move zfactors to ddm classes
