+++ MCEq/ddm_utils.py
@@ -453,154 +453,6 @@ def _gen_averaged_dndx(xbins: np.ndarray, entry) -> np.ndarray:
 
     return integral
 
-
-def calc_zfactor_and_error(
-    ddm, projectile: int, secondary: int, ebeam: Union[str, float], gamma: float = 1.7
-):
-    """
-    Calculate the Z-factor and its error for a given projectile,
-    secondary, and beam energy.
-
-    Parameters
-    ----------
-    ddm: DataDrivenModel
-        Instance of DataDrivenModel.
-    projectile : int
-        Projectile PDG code.
-    secondary : int
-        Secondary PDG code.
-    ebeam : float
-        Beam energy in GeV.
-    gamma : float, optional
-        CR nucleon integral spectral index. Default is 1.7.
-
-    Returns
-    -------
-    numpy.ndarray
-        Z-factor values.
-    numpy.ndarray
-        Error of the Z-factor values.
-
-    Notes
-    -----
-    The Z-factor is calculated by integrating the spline defined by the DDM entry
-    over the range from entry.x_min to 1.0. The integral is multiplied by the tuning
-    value of the entry and raised to the power of gamma.
-
-    The error of the Z-factor is propagated using the covariance matrix of the spline
-    and the provided _PROPAGATE_PARAMS.
-
-    Examples
-    --------
-    >>> model = DataDrivenModel(...)
-    >>> projectile = 2212
-    >>> secondary = 211
-    >>> ebeam = 10.0
-    >>> z_factor, z_error = model.calc_zfactor_and_error(projectile, secondary, ebeam)
-    """
-
-    entry = ddm.spline_db.get_entry(projectile, secondary, ebeam)
-
-    info(3, f"Calculating Z-factor for {projectile} --> {secondary} @ {ebeam} GeV.")
-
-    def func_int(tck_1):
-        res = quad(
-            _eval_spline,
-            entry.x_min,
-            1.0,
-            args=(
-                (entry.tck[0], tck_1, entry.tck[2]),
-                entry.x17,
-                entry.cov,
-                False,
-                gamma,
-            ),
-            **_QUAD_PARAMS,
-        )[0]
-        res = np.atleast_1d(res)
-        res[(res < 0) | ~np.isfinite(res) | (res > _LIMIT_PROPAGATE)] = 0.0
-        return res.squeeze()
-
-    y, C = propagate(func_int, entry.tck[1], entry.cov, **_PROPAGATE_PARAMS)
-    sig_y = np.sqrt(C)
-
-    return y, sig_y
-
-
-def calc_zfactor_and_error2(
-    ddm, projectile: int, secondary: int, ebeam: Union[str, float], gamma: float = 1.7
-):
-    """
-    Alternative Z-factor and error calculation for a given projectile,
-    secondary, and beam energy.
-
-    Parameters
-    ----------
-    ddm: DataDrivenModel
-        Instance of DataDrivenModel.
-    projectile : int
-        Projectile PDG code.
-    secondary : int
-        Secondary PDG code.
-    ebeam : float
-        Beam energy in GeV.
-    gamma : float, optional
-        CR nucleon integral spectral index. Default is 1.7.
-
-    Returns
-    -------
-    numpy.ndarray
-        Z-factor values.
-    numpy.ndarray
-        Error of the Z-factor values.
-
-    Notes
-    -----
-    The main difference to the previous function is the error calculation, which
-    is performed by integrating the error band of the evaluated splines and not
-    by propagating the errors of the integral. This method is a bit more stable
-    for the "problematic" splines.
-
-    Examples
-    --------
-    >>> model = DataDrivenModel(...)
-    >>> projectile = 2212
-    >>> secondary = 211
-    >>> ebeam = 158.0
-    >>> z_factor, z_error = model.calc_zfactor_and_error2(projectile, secondary, ebeam)
-    """
-
-    entry = ddm.spline_db.get_entry(projectile, secondary, ebeam)
-
-    info(3, f"Calculating Z-factor for {projectile}-->{secondary} @ {ebeam} GeV.")
-
-    def fitfunc_center(*args, **kwargs):
-        return _eval_spline(*args, **kwargs)[0]
-
-    def fitfunc_error(*args, **kwargs):
-        return _eval_spline(*args, **kwargs)[1]
-
-    zfactor_center = quad(
-        fitfunc_center,
-        entry.x_min,
-        1.0,
-        args=(entry.tck, entry.x17, entry.cov, True, gamma),
-        **_QUAD_PARAMS,
-    )[0]
-    zfactor_error = quad(
-        fitfunc_error,
-        entry.x_min,
-        1.0,
-        args=(entry.tck, entry.x17, entry.cov, True, gamma),
-        **_QUAD_PARAMS,
-    )[0]
-
-    return (
-        entry.tv * zfactor_center + entry.tv * entry.te * zfactor_error,
-        entry.te * entry.tv * zfactor_error,
-    )
-
-
 def gen_matrix_variations(ddm_obj, mceq):
     """
     Generate variations of DDM matrices and isospin partners.
